# Conflicts:
#	yarn-project/end-to-end/src/e2e_fees/fee_settings.test.ts
#	yarn-project/txe/src/state_machine/global_variable_builder.ts

+++ barretenberg/cpp/src/barretenberg/sumcheck/zk_sumcheck_data.hpp
@@ -151,7 +151,7 @@ template <typename Flavor> struct ZKSumcheckData {
         }
         total_sum *= scaling_factor;
 
-        return { total_sum + constant_term * (static_cast<uint64_t>(1) << libra_univariates.size()), scaling_factor };
+        return { total_sum + constant_term * (1UL << libra_univariates.size()), scaling_factor };
     }
 
     /**
